Added the ability to disable the logger
If you are using the logger in an instance type way against multiple machines, you might need to disable it for some reason.

+++ Functions/Common/New-LoggerObject/New-LoggerObject.ps1
@@ -71,6 +71,7 @@ Function New-LoggerObject {
     $loggerObject | Add-Member -MemberType NoteProperty -Name "NumberOfLogsToKeep" -Value $NumberOfLogsToKeep
     $loggerObject | Add-Member -MemberType NoteProperty -Name "WriteVerboseData" -Value $VerboseEnabled
     $loggerObject | Add-Member -MemberType NoteProperty -Name "PreventLogCleanup" -Value $false
+    $loggerObject | Add-Member -MemberType NoteProperty -Name "LoggerDisabled" -Value $false
     $loggerObject | Add-Member -MemberType ScriptMethod -Name "ToLog" -Value ${Function:Write-ToLog}
     $loggerObject | Add-Member -MemberType ScriptMethod -Name "WriteHostWriter" -Value ${Function:Write-ScriptMethodHostWriter}
     $loggerObject | Add-Member -MemberType ScriptMethod -Name "WriteVerboseWriter" -Value ${Function:Write-ScriptMethodVerboseWriter}
@@ -87,6 +88,10 @@ Function New-LoggerObject {
         param(
             [object]$LoggingString
         )
+
+        if ($this.LoggerDisabled) {
+            return
+        }
 
         if ($null -eq $LoggingString) {
             throw [System.Management.Automation.ParameterBindingException] "Failed to provide valid LoggingString"
@@ -106,6 +111,10 @@ Function New-LoggerObject {
             [object]$LoggingString
         )
 
+        if ($this.LoggerDisabled) {
+            return
+        }
+
         if ($null -eq $LoggingString) {
             throw [System.Management.Automation.ParameterBindingException] "Failed to provide valid LoggingString"
         }
@@ -123,6 +132,10 @@ Function New-LoggerObject {
             [object]$LoggingString
         )
 
+        if ($this.LoggerDisabled) {
+            return
+        }
+
         if ($null -eq $LoggingString) {
             throw [System.Management.Automation.ParameterBindingException] "Failed to provide valid LoggingString"
         }
@@ -135,6 +148,10 @@ Function New-LoggerObject {
     }
 
     $loggerObject | Add-Member -MemberType ScriptMethod -Name "UpdateFileLocation" -Value {
+
+        if ($this.LoggerDisabled) {
+            return
+        }
 
         if ($null -eq $this.FullPath) {
 
@@ -157,6 +174,10 @@ Function New-LoggerObject {
 
     $loggerObject | Add-Member -MemberType ScriptMethod -Name "LogUpKeep" -Value {
 
+        if ($this.LoggerDisabled) {
+            return
+        }
+
         if ($this.NextFileCheckTime -gt [System.DateTime]::Now) {
             return
         }
@@ -168,6 +189,10 @@ Function New-LoggerObject {
 
     $loggerObject | Add-Member -MemberType ScriptMethod -Name "CheckFileSize" -Value {
 
+        if ($this.LoggerDisabled) {
+            return
+        }
+
         $item = Get-ChildItem $this.FullPath
 
         if (($item.Length / 1MB) -gt $this.MaxFileSizeInMB) {
@@ -176,6 +201,10 @@ Function New-LoggerObject {
     }
 
     $loggerObject | Add-Member -MemberType ScriptMethod -Name "CheckNumberOfFiles" -Value {
+
+        if ($this.LoggerDisabled) {
+            return
+        }
 
         $filter = "{0}*" -f $this.InstanceBaseName
         $items = Get-ChildItem -Path $this.FileDirectory | Where-Object { $_.Name -like $filter }
@@ -190,10 +219,22 @@ Function New-LoggerObject {
 
     $loggerObject | Add-Member -MemberType ScriptMethod -Name "RemoveLatestLogFile" -Value {
 
+        if ($this.LoggerDisabled) {
+            return
+        }
+
         if (!$this.PreventLogCleanup) {
             $item = Get-ChildItem $this.FullPath
             Remove-Item $item -Force
         }
+    }
+
+    $loggerObject | Add-Member -MemberType ScriptMethod -Name "DisableLogger" -Value {
+        $this.LoggerDisabled = $true
+    }
+
+    $loggerObject | Add-Member -MemberType ScriptMethod -Name "EnableLogger" -Value {
+        $this.LoggerDisabled = $false
     }
 
     $loggerObject.UpdateFileLocation()
